+++ pkg/deploy/kong_plugin.go
@@ -26,8 +26,7 @@ func (k *K8sYaml) MountKongPlugin(functionName, pluginName, namespace, env strin
 		ApiVersion: "configuration.konghq.com/v1",
 		Kind:       "KongPlugin",
 		Metadata: metadata{
-			Name:      metadataName,
-			Namespace: namespace,
+			Name: metadataName,
 			Labels: map[string]string{
 				"global": stack.Functions[functionName].Plugins[pluginName].Global,
 			},